// bedrock-types.ts
export type ItemStackResponses = {
  /** Status specifies if the request with the RequestID below was successful. If this is the case, the ContainerInfo below will have information on what slots ended up changing. If not, the container info will be empty. A non-0 status means an error occurred and will result in the action being reverted. */
  status: "ok" | "error";
  /** RequestID is the unique ID of the request that this response is in reaction to. If rejected, the client will undo the actions from the request with this ID. */
  request_id: number;
  containers?: {
    slot_type: FullContainerName;
    slots: {
      slot: number;
      hotbar_slot: number;
      count: number;
      item_stack_id: number;
      custom_name: string;
      filtered_custom_name: string;
      durability_correction: number;
    }[];
  }[];
}[];
// ...
export type FullContainerName = {
  container_id: ContainerSlotType;
  dynamic_container_id?: number;
};
// ...
export type packet_item_stack_response = {
  /** Responses is a list of responses to ItemStackRequests sent by the client before. Responses either approve or reject a request from the client. Vanilla limits the size of this slice to 4096. */
  responses: ItemStackResponses;
};